# packages/contracts/openapi/dubhe-core.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dubhe Core",
    "description": "中文优先的 AI 投资研究与受控量化交易后端 API。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/devices/register": {
      "post": {
        "summary": "Register Device Endpoint",
        "operationId": "register_device_endpoint_v1_auth_devices_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/snapshot": {
      "get": {
        "summary": "Workspace Snapshot Endpoint",
        "operationId": "workspace_snapshot_endpoint_v1_workspaces__workspace_id__snapshot_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "since_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Since Sequence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/watchlist/{symbol}": {
      "put": {
        "summary": "Upsert Watchlist Item Endpoint",
        "operationId": "upsert_watchlist_item_endpoint_v1_workspaces__workspace_id__watchlist__symbol__put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/sync-events": {
      "get": {
        "summary": "List Sync Events Endpoint",
        "operationId": "list_sync_events_endpoint_v1_workspaces__workspace_id__sync_events_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "since_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Since Sequence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncEvent"
                  },
                  "title": "Response List Sync Events Endpoint V1 Workspaces  Workspace Id  Sync Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/news/analyze": {
      "post": {
        "summary": "Analyze News Endpoint",
        "operationId": "analyze_news_endpoint_v1_news_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/news/feed": {
      "get": {
        "summary": "News Feed Endpoint",
        "operationId": "news_feed_endpoint_v1_news_feed_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Market",
              "default": "US"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          },
          {
            "name": "live",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Live"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/strategy/spec/validate": {
      "post": {
        "summary": "Validate Strategy Endpoint",
        "operationId": "validate_strategy_endpoint_v1_strategy_spec_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategySpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/strategy/drafts/from-analysis": {
      "post": {
        "summary": "Draft Strategy From Analysis Endpoint",
        "operationId": "draft_strategy_from_analysis_endpoint_v1_strategy_drafts_from_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/backtests/replay": {
      "post": {
        "summary": "Run Replay Backtest Endpoint",
        "operationId": "run_replay_backtest_endpoint_v1_backtests_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/risk/evaluate": {
      "post": {
        "summary": "Evaluate Risk Endpoint",
        "operationId": "evaluate_risk_endpoint_v1_risk_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIntent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskDecision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/approvals": {
      "get": {
        "summary": "List Approval Requests Endpoint",
        "operationId": "list_approval_requests_endpoint_v1_approvals_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRequest"
                  },
                  "title": "Response List Approval Requests Endpoint V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{approval_id}/approve": {
      "post": {
        "summary": "Approve Request Endpoint",
        "operationId": "approve_request_endpoint_v1_approvals__approval_id__approve_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{approval_id}/reject": {
      "post": {
        "summary": "Reject Request Endpoint",
        "operationId": "reject_request_endpoint_v1_approvals__approval_id__reject_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/risk/kill-switch": {
      "get": {
        "summary": "Get Kill Switch Endpoint",
        "operationId": "get_kill_switch_endpoint_v1_risk_kill_switch_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchState"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set Kill Switch Endpoint",
        "operationId": "set_kill_switch_endpoint_v1_risk_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/simulation/paper-orders": {
      "get": {
        "summary": "List Paper Orders Endpoint",
        "operationId": "list_paper_orders_endpoint_v1_simulation_paper_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PaperOrder"
                  },
                  "type": "array",
                  "title": "Response List Paper Orders Endpoint V1 Simulation Paper Orders Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit Paper Order Endpoint",
        "operationId": "submit_paper_order_endpoint_v1_simulation_paper_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIntent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperOrder"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeviceRegistrationRequest": {
        "properties": {
          "account_key": {
            "type": "string",
            "minLength": 3,
            "title": "Account Key",
            "default": "local-demo"
          },
          "account_name": {
            "type": "string",
            "minLength": 1,
            "title": "Account Name",
            "default": "本地演示账户"
          },
          "device_name": {
            "type": "string",
            "minLength": 1,
            "title": "Device Name"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          }
        },
        "type": "object",
        "required": [
          "device_name",
          "platform"
        ],
        "title": "DeviceRegistrationRequest"
      },
      "DeviceSession": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          },
          "device_name": {
            "type": "string",
            "title": "Device Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "device_id",
          "workspace_id",
          "access_token",
          "platform",
          "device_name"
        ],
        "title": "DeviceSession"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "entity_type": {
            "$ref": "#/components/schemas/SyncEntityType"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "created",
              "updated",
              "deleted"
            ],
            "title": "Action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "sequence",
          "entity_type",
          "entity_id",
          "action"
        ],
        "title": "SyncEvent"
      },
...
      "WatchlistItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "minLength": 1,
            "title": "Workspace Id"
          },
          "symbol": {
            "type": "string",
            "minLength": 1,
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "notes_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes Zh"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "symbol",
          "name",
          "market"
        ],
        "title": "WatchlistItem"
      },
      "WatchlistUpsertRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "minLength": 1,
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "notes_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes Zh"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "market"
        ],
        "title": "WatchlistUpsertRequest"
      },
...
      "WorkspaceSnapshot": {
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/Workspace"
          },
          "watchlist": {
            "items": {
              "$ref": "#/components/schemas/WatchlistItem"
            },
            "type": "array",
            "title": "Watchlist"
          },
          "news_events": {
            "items": {
              "$ref": "#/components/schemas/NewsEvent"
            },
            "type": "array",
            "title": "News Events"
          },
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/NewsAnalysis"
            },
            "type": "array",
            "title": "Analyses"
          },
          "risk_decisions": {
            "items": {
              "$ref": "#/components/schemas/RiskDecision"
            },
            "type": "array",
            "title": "Risk Decisions"
          },
          "approval_requests": {
            "items": {
              "$ref": "#/components/schemas/ApprovalRequest"
            },
            "type": "array",
            "title": "Approval Requests"
          },
          "paper_orders": {
            "items": {
              "$ref": "#/components/schemas/PaperOrder"
            },
            "type": "array",
            "title": "Paper Orders"
          },
          "strategy_drafts": {
            "items": {
              "$ref": "#/components/schemas/StrategyDraft"
            },
            "type": "array",
            "title": "Strategy Drafts"
          },
          "backtest_results": {
            "items": {
              "$ref": "#/components/schemas/BacktestResult"
            },
            "type": "array",
            "title": "Backtest Results"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/SyncEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "server_sequence": {
            "type": "integer",
            "title": "Server Sequence",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "workspace"
        ],
        "title": "WorkspaceSnapshot"
      }
    }
  }
}